Results, Controls, Cleanup
I did results and Controls arhitecture
+I did some CSS cleanup
