+++ src/appComponents/Results.js
@@ -1,175 +1,28 @@
 'use client'
-import html2canvas from 'html2canvas';
 
-import {useRef, useState} from "react"
-import * as lcl from "../lcl"
+import {useRef} from "react"
+
+import OrderResultsCard from "./resultsBlocks/OrderResultsCard"
+import ResultInputs from "./resultsBlocks/ResultInputs"
 
 
-const Results = ({close, setClose, open, setOpen, init, _setInit}) => {
+/* 
+Result components provides results of an Order and Functionality to share it
 
-  const [shareSwitch, useShareSwitch] = useState(false)
-  const shareClick = () => {useShareSwitch(shareSwitch => !shareSwitch)}
-
-  let orderConfigParams = "";
-
-  orderConfigParams = `?am=${open.amount}&lev=${open.leverage}&op=${open.price}&fee=${open.fee}&long=${close.long}&cp=${close.price}&min=${close.min}&max=${close.max}`
-
-const getEncodedConfigParams = () => {return encodeURIComponent(`https://bepely.space/app${orderConfigParams}`)}
-const getDecodedConfigParams = (p) => {return decodeURIComponent(p)}
-
- const onCopy = () => {
-   navigator.clipboard.writeText(`https://bepely.space/app${orderConfigParams}`).then(() => {
-  
-  }, (err) => {
-  console.error('Could not copy text || ', err);
-  });
- }
+12.12.22 Bepely*/
+const Results = ({close, open, _setInit}) => {
 
     const snap = useRef()
 
-    const onSnap = async () => {
-      
-
-      const element = snap.current;
-      const canvas = await html2canvas(element, {backgroundColor: null});
-  
-      const data = canvas.toDataURL('image/png');
-      const link = document.createElement('a');
-
-     
-  
-      if (typeof link.download === 'string') {
-        
-        link.href = data;
-        link.download = `Leveroi ${close.long ? "Long" : "Short"}`;
-
-            document.body.appendChild(link);
-            link.click();
-            document.body.removeChild(link);
-          
-
-        
-      } else {
-        window.open(data);
-      }
-    }
-
-  const resultStages = [
-    "🤡","😰", "💥", "🔥", "⚡", "🌠",
-    "🎰",
-      "🌟", "🐗", "✨", "🌚", "🏆", "🤡" ]
-  let comment = resultStages[6]
-
-  const commentSelection = () => {
-    let asignee
-    
-
-    asignee = lcl.marginPercent(open, close)
-
-    if (asignee <= -1500)    {comment = resultStages[0]}
-    else if(asignee <= -500) {comment = resultStages[1]}
-    else if(asignee <= open.fee-100) {comment = resultStages[2]}
-    else if(asignee <= -80)  {comment = resultStages[3]}
-    else if(asignee <= -40)  {comment = resultStages[4]}
-    else if(asignee <= -5)   {comment = resultStages[5]}
-
-    else if(asignee <= 5 && asignee >= -5) {comment = resultStages[6]}
-
-    else if(asignee >= 1500) {comment = resultStages[12]}
-    else if(asignee >= 500)  {comment = resultStages[11]}
-    else if(asignee >= 100)  {comment = resultStages[10]}
-    else if(asignee >= 80)   {comment = resultStages[9]}
-    else if(asignee >= 40)   {comment = resultStages[8]}
-    else if(asignee >= 5)    {comment = resultStages[7]}
-    }
-
-  commentSelection()
-    
-
-  
- 
-
-  
-  
  return (
    
   <>
 
     
-  <div  className='layerFloor multiVer' id='' >
-        
-          
-          <div className="blockCard layerTable nested " id='resultsFloor' ref={snap}>
-            <div className="openResults exactResults" >
-             
-              <div className="orderData">
-              <div id="openInfo">
-                <h4>Bid:{open.amount * open.leverage}</h4>
-                <h5>Amount:{open.amount}</h5>
-                <h5>Leverage:{open.leverage}</h5>
-                <h5>{open.fee && open.fee !== 0 ? "Fee:" + open.fee + "%" : "No Fee"}</h5>
-              </div>
-                <div id="liquidation">
-                  <h5>💥 <br/> {lcl.liquidation(open, close)}</h5>
-                </div>
-            </div>
-              
-            
-              
-            </div>
-
-            <div className="closeResults exactResults">
-            <div id="orderTitle">
-               <p>{comment}</p>
-               <h3 id="resultsHeader">{close.long ? "Long Order" : "Short Order"}</h3>
-                 </div> 
-                <div className="pnlResults"> 
-                  <div className="pnlContainer" id='textContent'>
-                    <h4>PNL</h4>
-                    <h5>{lcl.margin(open, close)}</h5>
-                    <h5>{lcl.marginPercent(open, close)}%</h5>
-                  </div>
-                  <div className="totalContainer" id='textContent'>
-                    <h4>Total</h4>
-                    <h5>{lcl.total(open, close)}</h5>
-                  </div>
-                  
-                </div>
-            </div>
-            <div className="priceHolder" id="textContent">
-              <h5>Entry:{lcl.fixCoef(open.price)}</h5>
-              <h5>Close:{lcl.fixCoef(close.price)}</h5>
-            </div>
-          </div>
-        
-       </div>
-
-       <div className='layerBase nested multiVer' id='layerBaseFooterResults'>
-       
-          <div className='soloCenter layerTable' id='blockField'>
-            <input type="text" readOnly value={"https://bepely.space/app"+orderConfigParams}/>
-          </div>
-          {shareSwitch ? <div className='multiHor layerTable' id='blockButton3'>
-          <a  target="_blank" href={`https://t.me/share/url?url=${getEncodedConfigParams()} 
-&text=I am sharing ${close.long ? "Long" : "Short"} order simulation with ${lcl.margin(open, close)} profit || Made with Leveroi`
-        }> Telegram </a>  
-          
-            <a onClick={onSnap}> Make a Snap </a> 
-            <a onClick={onCopy}> Copy URL </a> 
-         
-          </div>
-          :
-          <></>
-          }
-        <div className="multiHor layerTable" id="blockButton2">          
-                
-          
-        <a onClick={shareClick}> {!shareSwitch ? "Share 👀" : "Hide Share"}</a>
-        <a onClick={_setInit}> Configurate new order</a>
-       </div>
-       </div>
+      <OrderResultsCard open={open} close={close} snap={snap}/>
+      <ResultInputs open={open} close={close} snap={snap} _setInit={_setInit}/>
      
-       </>
+  </>
                     
   )
