+++ src/styles/LDC/blocks.css
@@ -1,20 +1,18 @@
-#blockField{
+.blockField{
 width: 100%;
 min-height: 2vh;
 max-height: 5vh;
+border: none;
+background-color: #f8f8f8;
+overflow: hidden;
 
+height: 100%;
+
+
+
+border-radius: 4px;
 }
-#blockField>*{
 
-    border: none;
-    background-color: transparent;
-    overflow: hidden;
-    width: 100%;
-    height: 100%;
-    border: 2px solid black;
-
-    border-radius: 4px;
-}
 
 
 #blockButton2{grid-template-columns: 1fr 1fr}
@@ -68,7 +66,7 @@ gap: 10px;
 }
 
 .paddingContent{
-    padding: 20px 10px;
+    padding: 1vh 1vw;
 }
 
 
@@ -97,12 +95,10 @@ gap: 10px;
     #blockButton2, #blockButton3{
         width: 90%;
     }
-    #blockField{
+    .blockField{
         width: 90%;
     }
-    #blockField>*{
-        max-height: 4.5vw;
-    }
+    
     
     #blockButton2>*, #blockButton3>*{
         font-size: 1.6vmax;
